Refactor code and put types back in WebGoldenComparator
